refactor: split helpers, add Region type
